Bluetooth: controller: encryption and decryption using CAUv3 for Vega
Created unit tests for the encryption and decryption functions
Tested with the peripheral sample
Due to latency of CAUv3 when used as CCM inline accelerator
only one of the PDU can be encrypted/decrypted within an
bilateral exchange M->S + S->M in a connection event
If the RXed PDU is encrypted, the TXed PDU must be empty
with More Data if there is data in the queue. The TXed
PDU will be encrypted when an empty PDU is RXed.

Signed-off-by: Cristi Caciuloiu <cristian.caciuloiu@nxp.com>